remotepixel/requirements.txt
numpy
Pillow
mercantile
cachetools
rasterio[s3]==1.0a9
